+++ bindings/matrix-sdk-ffi/src/room/room_info.rs
@@ -14,7 +14,7 @@
 
 use std::sync::Arc;
 
-use matrix_sdk::{EncryptionState, RoomState};
+use matrix_sdk::{CallIntentConsensus, EncryptionState, RoomState};
 use tracing::warn;
 
 use crate::{
@@ -27,6 +27,34 @@ use crate::{
     room_member::RoomMember,
     ruma::RtcCallIntent,
 };
+
+#[derive(Clone, uniffi::Enum)]
+pub enum RtcCallIntentConsensus {
+    Full(RtcCallIntent),
+    Partial {
+        intent: RtcCallIntent,
+        agreeing_count: u64,
+        total_count: u64,
+    },
+    /// No consensus. No one advertises or advertisers disagree.
+    None,
+}
+
+impl From<CallIntentConsensus> for RtcCallIntentConsensus {
+    fn from(value: CallIntentConsensus) -> Self {
+        match value {
+            CallIntentConsensus::Full(intent) => RtcCallIntentConsensus::Full(intent.into()),
+            CallIntentConsensus::Partial { intent, agreeing_count, total_count } => {
+                RtcCallIntentConsensus::Partial {
+                    intent: intent.into(),
+                    agreeing_count,
+                    total_count,
+                }
+            }
+            CallIntentConsensus::None => RtcCallIntentConsensus::None,
+        }
+    }
+}
 
 #[derive(uniffi::Record)]
 pub struct RoomInfo {
@@ -70,7 +98,7 @@ pub struct RoomInfo {
     cached_user_defined_notification_mode: Option<RoomNotificationMode>,
     has_room_call: bool,
     active_room_call_participants: Vec<String>,
-    active_room_call_consensus_intent: Option<RtcCallIntent>,
+    active_room_call_consensus_intent: RtcCallIntentConsensus,
     /// Whether this room has been explicitly marked as unread
     is_marked_unread: bool,
     /// "Interesting" messages received in that room, independently of the
@@ -174,9 +202,7 @@ impl RoomInfo {
                 .iter()
                 .map(|u| u.to_string())
                 .collect(),
-            active_room_call_consensus_intent: room
-                .active_room_call_consensus_intent()
-                .map(Into::into),
+            active_room_call_consensus_intent: room.active_room_call_consensus_intent().into(),
             is_marked_unread: room.is_marked_unread(),
             num_unread_messages: room.num_unread_messages(),
             num_unread_notifications: room.num_unread_notifications(),